+++ pyTigerGraph/pytgasync/pyTigerGraphGSQL.py
@@ -14,7 +14,9 @@ from pyTigerGraph.common.exception import TigerGraphException
 from pyTigerGraph.common.gsql import (
     _parse_gsql,
     _prep_get_udf,
-    _parse_get_udf
+    _parse_get_udf,
+    _get_reserved_keywords,
+    _is_reserved_keyword,
 )
 
 from pyTigerGraph.pytgasync.pyTigerGraphBase import AsyncPyTigerGraphBase
@@ -411,7 +413,7 @@ class AsyncPyTigerGraphGSQL(AsyncPyTigerGraphBase):
 
         params = {}
         if verbose:
-            params["verbose"] = verbose
+            params["verbose"] = str(verbose).lower()
 
         res = await self._req("GET", self.gsUrl+"/gsql/v1/version",
                              params=params, authMode="pwd", resKey=None,
@@ -422,3 +424,24 @@ class AsyncPyTigerGraphGSQL(AsyncPyTigerGraphBase):
         logger.debug("exit: getGSQLVersion")
 
         return res
+
+    @staticmethod
+    def getReservedKeywords() -> frozenset:
+        """Return the full set of GSQL reserved keywords.
+
+        Returns:
+            A frozenset of uppercase keyword strings.
+        """
+        return _get_reserved_keywords()
+
+    @staticmethod
+    def isReservedKeyword(name: str) -> bool:
+        """Check whether *name* is a GSQL reserved keyword (case-insensitive).
+
+        Args:
+            name: The identifier to check.
+
+        Returns:
+            True if the name is reserved.
+        """
+        return _is_reserved_keyword(name)